vinyl: delete incomplete run files after recovery
If the server is shut down in the middle of a dump/compaction task,
there will be an incomplete run file left, i.e. a run that either was
not written to the end or was not inserted into a range. In either
case we are not going to use such a run file after restart and hence
need to delete it upon recovery completion. To achieve that, let's
write a VY_LOG_PREPARE_RUN record before writing a run file and collect
all runs that were "prepared", but not deleted or inserted into a range
in the vy_index_recovery::incomplete_runs. This way, we can clean them
up after recovery is complete.