vinyl: do not log run creation for dropped indexes
Records corresponding to dropped indexes are deleted from the metadata
log on rotation, so we must not log a run creation if the index this run
is for was dropped, otherwise we'll fail to recover with an unknown
index/range/run error. To avoid logging, call the ->abort() callback
instead of ->complete() on task completion if the index was dropped.

NOTE, this patch is rather a workaround committed for the sake of
testing. The issue addressed by it is fixed in a proper way by commit
"vinyl: delete incomplete run files after recovery", branch
gh-2029-vy-run-gc-v3, so this patch doesn't need to be applied if
gh-2029-vy-run-gc-v3 is merged first.