vinyl: remove 'first DELETE' optimization
We have been skipping the DELETE statements, which
were first DELETEs in an index, but such a case
is possible only once at index life time.
Also this optimization complicates the code and
intruduces additional overhead on each
vy_ranse_set_delete.

Another problem of this optimization is that we
in advance can't know will or will not the
DELETE statement be inserted, before calling
vy_range_set_delete.

To simplify the code and prepare to common vy_mem,
this optimization is removed.